jenkins: always test locale is up to date
Up till recently we had generated locales in the git and they were not
always kept up to date, but now we only keep .po file it is better to
always check they are up to date.

Always rebuild them and fail on make dist if missing.

(... and actually add the missing harmless update)

Change-Id: I8377fe4bca1b630ec56821f8d356efc54fadd0a7