scripts/version: fixup 'boot' versions 2020.4-atX -> 2020.4-at.X
mkswu 7.0 changed boot versions updates from different to higher, and
added a dot to make the version more semver-friendly (mixed alnum is
alpha order so at100 would be lower than at30)

Unfortunately, at.24 turned out to be considered lower than at20 (by
swupdate and busybox sort -V, but NOT by gnu sort -V...), so we need
to be extra careful with these versions:
- for new versions, always do the substitution
- for old versions, always replace but if this would have been an
upgrade and is not one then keep sw-versions in the old value.

It is possible that the "real" new version was also in old format in
which case the update would have happened, but the worst that can happen
is we'd allow updating the same uboot twice, so we do not need to care
about the non-fixed new version.

Link: https://armadillo.atmark-techno.com/forum/armadillo/24108
Change-Id: I53280d82ca516bc74c4f6d19af426671a0658ecb