scripts/pre_appfs.sh: fix not to use podman stop when killing container
podman_killall() is used to stop containers while executing SWUpdate.
When executing SWUpdate via USB (in unshare -m), and update failed with
CONTAINER_CLEAR due to `podman stop -a` never exiting.

This is because podman stop tries to remove the network namespace
indefinitely from 5.2.1/5.3.0 and thus never terminates.
(Note podman 5.4.0 introduced a 60s timeout here, but this is still much
slower than required)

To resolve this, modify the process to stop containers by sending
sigterm/sigkill directly to the container as appropriate.

Note this will also be resolved when we convert swupdate to run as a
service (even if swupdate-usb runs in a different mount namespace the
actual install process will happen in main namespace), but there is no
downside to using kill; the reason we used stop in the first place was
to send a sigterm first but we can do this manually.

Fixes: 303f88fbf2be ("scripts/appfs: use podman stop instead of podman kill")
Change-Id: I7b24fef1136f70474c5972d9608e353ceb08b30f