scripts/podman_cleanup: do not try to delete an image multiple times
podman image list will list the same image multiple times if it has
multiple tags, which if the image is not used would lead to multiple
podman rm where the later ones will fail.

We ignore errors so this is not too horrible, but remove duplicates
early on to avoid this.

Note that podman_cleanup leaves unused tags if any of the tag of the
image is used, if that becomes a problem in the future we might want
to update the algorithm to consider tags directly...

Change-Id: I8f586bef8760932cf7566401c3bc111718386e1d