scripts/post_rootfs: check dtb and dtbos are present
Check for dtb and dtbos when updating rootfs.

uboot will silently skip DTBOs listen in overlays.txt that could not be
read, so users might not notice if their dtbo is gone after upgrade

It is now an error if a freshly removed file is still listed in
overlays.txt, or if a file newly listed to overlays.txt is missing

If the file had already been missing, we cannot do anything about it now
so it is only a warning.

Change-Id: I8a8e5478316ce82eccf2e446187318460c06a570