swdesc_command/exec: ensure logic within commands is isolated
swdesc_command and swdesc_exec link successive arguments with '&&', but
if commands use logic operators such as '||' this can have unwanted impact
e.g.
    swdesc_command false 'true || echo should not run'
will be translated to
    false && true || echo should not run
and will run the echo

This makes the command to be transformed to
   { false; } && { true || echo should not run; }
and will properly stop at the first false, and not interpret that or as
an escape hatch.

Change-Id: I0fdec65047b7eaf84878202fc542615f1d108f86