scripts/pre_init: unset upgrade_available with '00' instead of unset
When using secureboot with env_writeable_list, unsetting the variable is not
convenient because that would fallback to the default value.

This change comes with a change in ABOS to handle the '00' value properly,
but even without the ABOS change this does not break anything other than
the status message if an update fails halfway, and rollback is properly
disabled with 00 just as it is with 0.
It still displays a warning on boot though so add a fallback for it for
a while, it will be removed at some point next year

Change-Id: I6d61a0e7ef9f5a27132b41910d6f0cc6530b89ce