scripts/pre_boot: avoid copying zeroed bootloader around
This is a workaround for a potential bug on A900, at least until we
understand what happened, but there is no hurt in making sure:

Always check the first 4k of boot partition are not zeroes.
If it's zeroes:
- if backup partition is also dead, ensure bootloader can be reinstalled
and fail update with a warning
- if backup partition is present, recover from that partition but also
mark boot version as uninstalled to allow clean reinstall.

Also print a warning, but in practice this is not really useful as there
is nothing we can do about it unless this can be reproduced.
It would be interesting to have a mode that fails update completely even
if B side is acceptable, to allow for investigation instead, but that
can come later.

Change-Id: I75ee66d1d8d15c7e2acad0a3c8943f4002e3d1cb