mkswu --show: avoid printing rootfs twice
mkswu now creates a second dummy file to trigger post rootfs scripts,
which was incorrectly listed as rootfs being ran twice in mkswu --show
output.

This is not straightforward to fix, so add an extra comment prop to hide
anything from show and be done with it.
That means older swu will still list base os twice, this is fine as the
show content is only meant to be informative and not faithful

Change-Id: I48c57833b7225444ea2d9ddc9ed71a47b12c3982