scripts/rootfs: preserve files: reset files owners to root
When copying files from one rootfs to the next, the uid for non-root
users is not stable, which was why we added a CHOWN directive to
swupdate_preserve_files that runs chown in the /target chroot.

Up until mkswu 7.8 this was enough because all the files copied in the
default list were root-owned, except for the abos-web directory which
was chowned afterward, but with the v15 preserve list we're now copying
/etc/atmark/abos_web without chown and when upgrading from an older
system which was abos-web-admin owned then the uid is incorrectly
preserved across the copy.

Coreutils cp would allow specifying which attributes to preserve as a
list to not copy owner, but busybox does not allow preserving mode and
timestamps without also copying the owner, so run an extra find+chown
pass after copy to reset all copied files to root:root as originally
intended.

This fixes leftover files in /etc/atmark/abos_web directory that have a
potentially incorrect owner.

Change-Id: Id884f5fe7889b1b548f583b600c8d6ec6da5d40d